requirements.txt
# Requerimientos principales para hestia-utils-python
pyyaml
PyYAML>=6.0
